# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ScanRequest {
  /** A URL (e.g. https://example.com) or a raw HTTP request (with a Host header) */
  input: string;
}

export interface OriginVerifyRequest {
  /** The original target hostname to send as Host header / TLS SNI */
  hostname: string;
  /** The candidate origin IP to connect to directly */
  ip: string;
  /** Port to connect on */
  port?: number;
  /** Whether to use TLS (HTTPS) for the direct connection */
  useHttps?: boolean;
}

/**
 * Best-effort classification. "likely_origin" means multiple independent signals agree; it is not proof of internal network topology, which cannot be fully confirmed from outside.
 */
export type OriginVerifyResultVerdict = typeof OriginVerifyResultVerdict[keyof typeof OriginVerifyResultVerdict];


export const OriginVerifyResultVerdict = {
  likely_origin: 'likely_origin',
  possible_proxy: 'possible_proxy',
  indeterminate: 'indeterminate',
} as const;

export interface OriginVerifyResult {
  hostname: string;
  ip: string;
  reachable: boolean;
  statusCode?: number | null;
  statusText?: string | null;
  server?: string | null;
  responseTimeMs?: number | null;
  tlsCertMatchesHost?: boolean | null;
  tlsCertSubject?: string | null;
  tlsCertIssuer?: string | null;
  bodyPreview?: string | null;
  /** Header names found on the direct response that typically indicate an intermediary proxy/load-balancer hop (Via, X-Cache, X-Served-By, X-Forwarded-*, X-Real-IP) */
  proxyHeadersDetected: string[];
  /** HTTP status returned by the normal public hostname (via DNS/CDN) for comparison */
  publicStatusCode?: number | null;
  /** 0-1 trigram similarity between the direct-IP response body and the public hostname's response body */
  publicBodySimilarity?: number | null;
  /** 0-1 overlap ratio between the response header sets of the direct-IP response and the public hostname response */
  publicHeaderOverlap?: number | null;
  /** Best-effort classification. "likely_origin" means multiple independent signals agree; it is not proof of internal network topology, which cannot be fully confirmed from outside. */
  verdict: OriginVerifyResultVerdict;
  verdictReason?: string | null;
  error?: string | null;
}

export interface ScanError {
  message: string;
}

export type ProxyRequestBodyMethod = typeof ProxyRequestBodyMethod[keyof typeof ProxyRequestBodyMethod];


export const ProxyRequestBodyMethod = {
  GET: 'GET',
  POST: 'POST',
  PUT: 'PUT',
  PATCH: 'PATCH',
  DELETE: 'DELETE',
  HEAD: 'HEAD',
  OPTIONS: 'OPTIONS',
} as const;

/**
 * Extra request headers to send (Host is always overridden with hostname)
 */
export type ProxyRequestBodyHeaders = {[key: string]: string};

export interface ProxyRequestBody {
  /** The original target hostname to send as Host header / TLS SNI */
  hostname: string;
  /** The candidate origin IP to connect to directly */
  ip: string;
  /** Port to connect on */
  port?: number;
  /** Whether to use TLS (HTTPS) for the direct connection */
  useHttps?: boolean;
  method: ProxyRequestBodyMethod;
  /** Request path (and query string), e.g. /api/users?id=1 */
  path: string;
  /** Extra request headers to send (Host is always overridden with hostname) */
  headers?: ProxyRequestBodyHeaders;
  /** Raw request body to send (ignored for GET/HEAD) */
  body?: string | null;
}

export type ProxyRequestResultHeaders = {[key: string]: string};

export interface ProxyRequestResult {
  reachable: boolean;
  statusCode?: number | null;
  statusText?: string | null;
  responseTimeMs?: number | null;
  headers: ProxyRequestResultHeaders;
  body?: string | null;
  bodyTruncated: boolean;
  error?: string | null;
}

export interface DnsResolverResult {
  resolver: string;
  addresses: string[];
  error?: string | null;
}

export interface SslCertificateInfo {
  subject?: string | null;
  issuer?: string | null;
  validFrom?: string | null;
  validTo?: string | null;
  altNames?: string[];
  fingerprint?: string | null;
  error?: string | null;
}

export interface SubdomainRecord {
  hostname: string;
  addresses: string[];
}

export type CandidateOriginIpConfidence = typeof CandidateOriginIpConfidence[keyof typeof CandidateOriginIpConfidence];


export const CandidateOriginIpConfidence = {
  high: 'high',
  medium: 'medium',
  low: 'low',
} as const;

export interface CandidateOriginIp {
  ip: string;
  confidence: CandidateOriginIpConfidence;
  sources: string[];
  /** ISP / hosting organization that owns this IP, from ASN lookup */
  org?: string | null;
  /** City/country the IP geolocates to */
  location?: string | null;
}

export interface IpDetails {
  ip: string;
  org?: string | null;
  location?: string | null;
}

export interface ScanResult {
  originalInput: string;
  hostname: string;
  cdnDetected: boolean;
  cdnProvider: string | null;
  edgeIps: string[];
  dnsResults: DnsResolverResult[];
  sslCertificate: SslCertificateInfo;
  mxRecords: string[];
  spfRecord: string | null;
  txtRecords: string[];
  subdomains: SubdomainRecord[];
  candidateOriginIps: CandidateOriginIp[];
  edgeIpDetails: IpDetails[];
  scannedAt: string;
}