# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: recon
    description: Reverse proxy / origin IP detection operations
paths:
  /scan:
    post:
      operationId: scanTarget
      tags: [recon]
      summary: Analyze a URL or raw HTTP request for the real origin server behind a CDN/edge proxy
      description: >
        Accepts a URL or raw HTTP request, extracts the hostname, then queries multiple
        DNS-over-HTTPS resolvers, inspects the live SSL certificate, checks MX/SPF/TXT
        records, and searches certificate-transparency logs for subdomains that may
        expose the real origin IP behind a CDN such as Cloudflare, Akamai, or Fastly.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ScanRequest"
      responses:
        "200":
          description: Scan result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ScanResult"
        "400":
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ScanError"
  /verify-origin:
    post:
      operationId: verifyOrigin
      tags: [recon]
      summary: Attempt to directly reach a candidate origin IP using the target's Host header
      description: >
        Connects directly to a candidate origin IP (bypassing DNS/CDN routing) and sends
        an HTTP request with the target hostname in the Host header and TLS SNI. Reports
        whether the origin responds, its status code, response headers, and whether its
        TLS certificate matches the target hostname — proving the CDN can be bypassed.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OriginVerifyRequest"
      responses:
        "200":
          description: Verification result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OriginVerifyResult"
        "400":
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ScanError"
  /proxy-request:
    post:
      operationId: sendProxyRequest
      tags: [recon]
      summary: Send a custom HTTP request directly to a candidate origin IP
      description: >
        Connects directly to a candidate origin IP (bypassing DNS/CDN routing) and sends
        a fully custom HTTP request — method, path, headers, and body — with the target
        hostname spoofed in the Host header and TLS SNI. Returns the raw response from
        the backend so you can interact with it as if the CDN/proxy were not there.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProxyRequestBody"
      responses:
        "200":
          description: Response from the origin server
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProxyRequestResult"
        "400":
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ScanError"
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"
components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status
    ScanRequest:
      type: object
      properties:
        input:
          type: string
          description: A URL (e.g. https://example.com) or a raw HTTP request (with a Host header)
      required:
        - input
    OriginVerifyRequest:
      type: object
      properties:
        hostname:
          type: string
          description: The original target hostname to send as Host header / TLS SNI
        ip:
          type: string
          description: The candidate origin IP to connect to directly
        port:
          type: integer
          description: Port to connect on
          default: 443
        useHttps:
          type: boolean
          description: Whether to use TLS (HTTPS) for the direct connection
          default: true
      required:
        - hostname
        - ip
    OriginVerifyResult:
      type: object
      properties:
        hostname:
          type: string
        ip:
          type: string
        reachable:
          type: boolean
        statusCode:
          type: integer
          nullable: true
        statusText:
          type: string
          nullable: true
        server:
          type: string
          nullable: true
        responseTimeMs:
          type: integer
          nullable: true
        tlsCertMatchesHost:
          type: boolean
          nullable: true
        tlsCertSubject:
          type: string
          nullable: true
        tlsCertIssuer:
          type: string
          nullable: true
        bodyPreview:
          type: string
          nullable: true
        proxyHeadersDetected:
          type: array
          items:
            type: string
          description: >
            Header names found on the direct response that typically indicate an
            intermediary proxy/load-balancer hop (Via, X-Cache, X-Served-By, X-Forwarded-*, X-Real-IP)
        publicStatusCode:
          type: integer
          nullable: true
          description: HTTP status returned by the normal public hostname (via DNS/CDN) for comparison
        publicBodySimilarity:
          type: number
          nullable: true
          description: 0-1 trigram similarity between the direct-IP response body and the public hostname's response body
        publicHeaderOverlap:
          type: number
          nullable: true
          description: 0-1 overlap ratio between the response header sets of the direct-IP response and the public hostname response
        verdict:
          type: string
          enum: [likely_origin, possible_proxy, indeterminate]
          description: >
            Best-effort classification. "likely_origin" means multiple independent signals agree;
            it is not proof of internal network topology, which cannot be fully confirmed from outside.
        verdictReason:
          type: string
          nullable: true
        error:
          type: string
          nullable: true
      required:
        - hostname
        - ip
        - reachable
        - proxyHeadersDetected
        - verdict
    ScanError:
      type: object
      properties:
        message:
          type: string
      required:
        - message
    ProxyRequestBody:
      type: object
      properties:
        hostname:
          type: string
          description: The original target hostname to send as Host header / TLS SNI
        ip:
          type: string
          description: The candidate origin IP to connect to directly
        port:
          type: integer
          description: Port to connect on
          default: 443
        useHttps:
          type: boolean
          description: Whether to use TLS (HTTPS) for the direct connection
          default: true
        method:
          type: string
          enum: [GET, POST, PUT, PATCH, DELETE, HEAD, OPTIONS]
          default: GET
        path:
          type: string
          description: Request path (and query string), e.g. /api/users?id=1
          default: /
        headers:
          type: object
          additionalProperties:
            type: string
          description: Extra request headers to send (Host is always overridden with hostname)
        body:
          type: string
          nullable: true
          description: Raw request body to send (ignored for GET/HEAD)
      required:
        - hostname
        - ip
        - method
        - path
    ProxyRequestResult:
      type: object
      properties:
        reachable:
          type: boolean
        statusCode:
          type: integer
          nullable: true
        statusText:
          type: string
          nullable: true
        responseTimeMs:
          type: integer
          nullable: true
        headers:
          type: object
          additionalProperties:
            type: string
        body:
          type: string
          nullable: true
        bodyTruncated:
          type: boolean
        error:
          type: string
          nullable: true
      required:
        - reachable
        - headers
        - bodyTruncated
    DnsResolverResult:
      type: object
      properties:
        resolver:
          type: string
        addresses:
          type: array
          items:
            type: string
        error:
          type: string
          nullable: true
      required:
        - resolver
        - addresses
    SslCertificateInfo:
      type: object
      properties:
        subject:
          type: string
          nullable: true
        issuer:
          type: string
          nullable: true
        validFrom:
          type: string
          nullable: true
        validTo:
          type: string
          nullable: true
        altNames:
          type: array
          items:
            type: string
        fingerprint:
          type: string
          nullable: true
        error:
          type: string
          nullable: true
    SubdomainRecord:
      type: object
      properties:
        hostname:
          type: string
        addresses:
          type: array
          items:
            type: string
      required:
        - hostname
        - addresses
    CandidateOriginIp:
      type: object
      properties:
        ip:
          type: string
        confidence:
          type: string
          enum: [high, medium, low]
        sources:
          type: array
          items:
            type: string
        org:
          type: string
          nullable: true
          description: ISP / hosting organization that owns this IP, from ASN lookup
        location:
          type: string
          nullable: true
          description: City/country the IP geolocates to
      required:
        - ip
        - confidence
        - sources
    IpDetails:
      type: object
      properties:
        ip:
          type: string
        org:
          type: string
          nullable: true
        location:
          type: string
          nullable: true
      required:
        - ip
    ScanResult:
      type: object
      properties:
        originalInput:
          type: string
        hostname:
          type: string
        cdnDetected:
          type: boolean
        cdnProvider:
          type: string
          nullable: true
        edgeIps:
          type: array
          items:
            type: string
        dnsResults:
          type: array
          items:
            $ref: "#/components/schemas/DnsResolverResult"
        sslCertificate:
          $ref: "#/components/schemas/SslCertificateInfo"
        mxRecords:
          type: array
          items:
            type: string
        spfRecord:
          type: string
          nullable: true
        txtRecords:
          type: array
          items:
            type: string
        subdomains:
          type: array
          items:
            $ref: "#/components/schemas/SubdomainRecord"
        candidateOriginIps:
          type: array
          items:
            $ref: "#/components/schemas/CandidateOriginIp"
        edgeIpDetails:
          type: array
          items:
            $ref: "#/components/schemas/IpDetails"
        scannedAt:
          type: string
      required:
        - originalInput
        - hostname
        - cdnDetected
        - cdnProvider
        - edgeIps
        - dnsResults
        - sslCertificate
        - mxRecords
        - spfRecord
        - txtRecords
        - subdomains
        - candidateOriginIps
        - edgeIpDetails
        - scannedAt