# lib/api-zod/src/generated/types/proxyRequestBody.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ProxyRequestBodyHeaders } from './proxyRequestBodyHeaders';
import type { ProxyRequestBodyMethod } from './proxyRequestBodyMethod';

export interface ProxyRequestBody {
  /** The original target hostname to send as Host header / TLS SNI */
  hostname: string;
  /** The candidate origin IP to connect to directly */
  ip: string;
  /** Port to connect on */
  port?: number;
  /** Whether to use TLS (HTTPS) for the direct connection */
  useHttps?: boolean;
  method: ProxyRequestBodyMethod;
  /** Request path (and query string), e.g. /api/users?id=1 */
  path: string;
  /** Extra request headers to send (Host is always overridden with hostname) */
  headers?: ProxyRequestBodyHeaders;
  /** Raw request body to send (ignored for GET/HEAD) */
  body?: string | null;
}