# lib/api-zod/src/generated/types/proxyRequestResult.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ProxyRequestResultHeaders } from './proxyRequestResultHeaders';

export interface ProxyRequestResult {
  reachable: boolean;
  statusCode?: number | null;
  statusText?: string | null;
  responseTimeMs?: number | null;
  headers: ProxyRequestResultHeaders;
  body?: string | null;
  bodyTruncated: boolean;
  error?: string | null;
}